Create a separate file for ScanCard
